Move UserType enum to auth types, only let Paras see Assigned link
