+++ src/backend/routers/para.test.ts
@@ -1,6 +1,6 @@
 import test from "ava";
 import { getTestServer } from "@/backend/tests";
-import { UserType } from "@/types/global";
+import { UserType } from "@/types/auth";
 
 test("getParaById", async (t) => {
   const { trpc, db } = await getTestServer(t, {
