+++ src/types/global.ts
@@ -4,9 +4,3 @@ export type Goal = SelectableForTable<"goal">;
 export type Subgoal = SelectableForTable<"subgoal">;
 export type ChangeEvent = React.ChangeEvent<HTMLInputElement>;
 export type FormEvent = React.FormEvent<HTMLFormElement>;
-export enum UserType {
-  User = "user",
-  Para = "para",
-  CaseManager = "case_manager",
-  Admin = "admin",
-}
